test(agent-gate-eval): move unit tests out of src/ into test/
Review 4832542711: test files should not live under src/. loading.test.ts
and verdict.test.ts move to internal/agent-gate-eval/test/ (imports repointed
to ../src/, tsconfig include extended). vitest discovers them by default glob;
36/36 still pass.
